Add comments; improve ease of log migration; improve example log config
