+++ gemini-resin-archetype/src/main/resources/archetype-resources/pom.xml
@@ -51,6 +51,11 @@
       <artifactId>gemini-jdbc</artifactId>
       <version>${geminiVersion}</version>
     </dependency>
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-api</artifactId>
+      <version>1.8.0-beta4</version>
+    </dependency>
     <!-- gemini-hikaricp is an alternative to gemini-jdbc -->
     <!-- <dependency> -->
     <!-- <groupId>com.techempower</groupId> -->
