+++ gemini/pom.xml
@@ -111,6 +111,7 @@
     <dependency>
       <groupId>org.slf4j</groupId>
       <artifactId>slf4j-api</artifactId>
+      <version>1.8.0-beta4</version>
     </dependency>
   </dependencies>
